requirements.txt
# BOT DEPENDENCIES #

praw				# base package to run
requests			# http requests
PILLOW				# image manipulation
gfycat              #for gfycat
opencv-python-headless		# video manipulaton
bs4					# http parsing
timeloop			#comment downvote checking

# HEROKU DEPENDENCIES #
airbrake			# for error logging
keen				# analytics
redis				# memory
